requirements.txt
backtrader
pandas
numpy
matplotlib
#alpaca-trade-api
alpaca-py
vectorbt
yfinance